Fix issues caused by missing null terminator in property_getTypeName
property_getTypeName in JastorRuntimeHelper.m creates a NSData object from a character buffer, but that buffer was not null-terminated. This caused the subsequent NSString::stringWithUTF8String to sometimes fail, which resulted in uninitialized objects, or worse.

Strangely, it failed only if the property name was 16 characters long; probably some memory alignment luck saved it in other cases.